Consolidate CreateCategoryPane and EditCategoryPane into one
CreateCategoryPane and EditCategoryPane have very similar logic and layout. They have been combined into a single ManageCategoryPane, which fulfills the functions of both. Pass ManageCategoryPane a category, and it will edit this category; pass it null, and it will create a new category.
